+++ src/client/components/Records/SubscriptionsRecords.tsx
@@ -1,5 +1,7 @@
 // Import Dependencies
 import React, { useState, ChangeEvent, useEffect } from 'react';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
@@ -8,17 +10,17 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import Input from "@material-ui/core/Input";
+import Input from '@material-ui/core/Input';
 import EditIcon from '@mui/icons-material/Edit';
 import DoneIcon from '@mui/icons-material/Done';
 import DeleteIcon from '@mui/icons-material/Delete';
+import swal from 'sweetalert';
 import axios from 'axios';
 
 interface Column {
   id: 'id' | 'season' | 'year' | 'flat_price' | 'weekly_price' | 'ActionDescription' 
   label: string;
   minWidth?: number;
-  align?: 'right';
   format?: (value: number) => string;
 }
 
@@ -48,6 +50,7 @@ const SubscriptionsRecords = () => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [rows, setRows] = useState([]);
+  const [rowEditing, setRowEditing] = useState(null);
   const [editing, setEditing] = useState(false);
   const [rowColumnId, setRowColumnId] = useState({});
   const [deleteCount, setDeleteCount] = useState(0);
@@ -66,6 +69,15 @@ const SubscriptionsRecords = () => {
   const patchSubscription = async (subscriptionsId: string, updatedSubscription: any) => {
     try {
       const { data } = await axios.patch(`/api/subscriptions/${subscriptionsId}`, updatedSubscription);
+      toast.success('Vendor Updated', {
+        position: "top-right",
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        });
       return data
     } catch (err) {
       console.error(err)
@@ -76,15 +88,31 @@ const SubscriptionsRecords = () => {
   }
   
   const deleteSubscription = async (subscriptionsId: string) => {
-    try {
-      const { data } = await axios.delete(`/api/subscriptions/${subscriptionsId}`);
-      return data
-    } catch (err) {
-      console.error(err)
-      return {
-        error: err
+    swal({
+      title: 'Are you sure?',
+      text: 'Vendor will be deleted, along with all associated products!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then( async (willDelete) => {
+      if (willDelete) {
+        swal('Event has been deleted', {
+          icon: 'success',
+        });
+        try {
+          const { data } = await axios.delete(`/api/subscriptions/${subscriptionsId}`);
+          setDeleteCount((deleteCount) => deleteCount + 1)
+          return data;
+        } catch (err) {
+          console.error(err);
+          return {
+            error: err,
+          };
+        }
+      } else {
+        swal('That was a close one!');
       }
-    }
+    });
   }
 
   useEffect(() => {
@@ -123,17 +151,29 @@ const SubscriptionsRecords = () => {
     patchSubscription(row.id, row);
   }
 
-  const onEdit = () => {
+  const onEdit = (row: object) => {
     setEditing(!editing)
+    setRowEditing(row.id)
   }
 
   const onDelete = (row: object) => {
-    setDeleteCount(deleteCount + 1)
     deleteSubscription(row.id)
+    setDeleteCount(deleteCount + 1)
   }
 
   return (
-    <Paper sx={{ width: '100%', overflow: 'hidden' }}>
+    <Paper sx={{ width: '90%', overflow: 'hidden' }}>
+      <ToastContainer
+        position="top-right"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+      />
       <TableContainer sx={{ maxHeight: 440 }}>
         <Table stickyHeader aria-label='sticky table'>
           <TableHead>
@@ -158,7 +198,7 @@ const SubscriptionsRecords = () => {
                     const value = row[column.id];
                     return (
                       <TableCell key={column.id} align={column.align}>
-                        {editing ? (
+                        {(editing && row.id === rowEditing) ? (
                           <Input
                             // type={String}
                             defaultValue={value}
@@ -172,18 +212,18 @@ const SubscriptionsRecords = () => {
                       </TableCell>
                     );
                   })}
-                  {editing ? (
+                  {(editing && row.id === rowEditing) ? (
                     <DoneIcon onClick={() => onDone(row)} />
-                  ) :
+                  ) : (
                     <>
                       <TableCell>
-                        <EditIcon onClick={() => onEdit()} />
+                        <EditIcon onClick={() => onEdit(row)} />
                       </TableCell>
                       <TableCell>
                         <DeleteIcon onClick={() => onDelete(row)} />
                       </TableCell>
                     </>
-                  }
+                  )}
 
                 </TableRow>
               ))}
