+++ src/client/components/Records/UsersRecords.tsx
@@ -1,5 +1,7 @@
 // Import Dependencies
 import React, { useState, ChangeEvent, useEffect } from 'react';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import Paper from '@mui/material/Paper';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
@@ -8,24 +10,24 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import Input from "@material-ui/core/Input";
+import Input from '@material-ui/core/Input';
 import EditIcon from '@mui/icons-material/Edit';
 import DoneIcon from '@mui/icons-material/Done';
 import DeleteIcon from '@mui/icons-material/Delete';
+import swal from 'sweetalert';
 import axios from 'axios';
 
 interface Column {
   id:
-  | 'id'
-  | 'googleId'
-  | 'name'
-  | 'email'
-  | 'street_address'
-  | 'roleId'
-  | 'delivery_zone';
+    | 'id'
+    | 'googleId'
+    | 'name'
+    | 'email'
+    | 'street_address'
+    | 'roleId'
+    | 'delivery_zone';
   label: string;
   minWidth?: number;
-  align?: 'right';
   format?: (value: number) => string;
 }
 
@@ -64,12 +66,14 @@ const UsersRecords = () => {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [rows, setRows] = useState([]);
+  const [rowEditing, setRowEditing] = useState(null);
   const [editing, setEditing] = useState(false);
   const [rowColumnId, setRowColumnId] = useState({});
   const [deleteCount, setDeleteCount] = useState(0);
 
   const getUsers = () => {
-    axios.get('/api/users')
+    axios
+      .get('/api/users')
       .then((data) => {
         // console.log(data.data);
         setRows(data.data);
@@ -82,29 +86,54 @@ const UsersRecords = () => {
   const patchUsers = async (userId: string, updatedUser: any) => {
     try {
       const { data } = await axios.patch(`/api/users/${userId}`, updatedUser);
-      return data
-    } catch (err) {
-      console.error(err)
-      return {
-        error: err
-      }
-    }
-  }
-
-  const deleteUser = async (userId: string) => {
-    try {
-      const {data} = await axios.delete(`/api/users/${userId}`);
+      toast.success('Vendor Updated', {
+        position: 'top-right',
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+      });
       return data;
     } catch (err) {
-      console.error(err)
+      console.error(err);
       return {
-        error: err
-      }
+        error: err,
+      };
     }
-  }
+  };
+
+  const deleteUser = async (userId: string) => {
+    swal({
+      title: 'Are you sure?',
+      text: 'Vendor will be deleted, along with all associated products!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        swal('Event has been deleted', {
+          icon: 'success',
+        });
+        try {
+          const { data } = await axios.delete(`/api/users/${userId}`);
+          setDeleteCount((deleteCount) => deleteCount + 1);
+          return data;
+        } catch (err) {
+          console.error(err);
+          return {
+            error: err,
+          };
+        }
+      } else {
+        swal('That was a close one!');
+      }
+    });
+  };
 
   useEffect(() => {
-    getUsers()
+    getUsers();
   }, [deleteCount]);
 
   const handleChangePage = (event: unknown, newPage: number) => {
@@ -132,24 +161,36 @@ const UsersRecords = () => {
     // console.log();
     setRows(newRows);
   };
-  
+
   const onDone = (row: object) => {
     // console.log(row)
-    setEditing(!editing)
+    setEditing(!editing);
     patchUsers(row.id, row);
-  }
+  };
 
-  const onEdit = () => {
-    setEditing(!editing)
-  }
+  const onEdit = (row: object) => {
+    setEditing(!editing);
+    setRowEditing(row.id);
+  };
 
   const onDelete = (row: object) => {
-    setDeleteCount(deleteCount + 1)
-    deleteUser(row.id)
-  }
+    deleteUser(row.id);
+    setDeleteCount(deleteCount + 1);
+  };
 
   return (
-    <Paper sx={{ width: '100%', overflow: 'hidden' }}>
+    <Paper sx={{ width: '90%', overflow: 'hidden' }}>
+      <ToastContainer
+        position='top-right'
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+      />
       <TableContainer sx={{ maxHeight: 440 }}>
         <Table stickyHeader aria-label='sticky table'>
           <TableHead>
@@ -174,33 +215,33 @@ const UsersRecords = () => {
                     const value = row[column.id];
                     return (
                       <TableCell key={column.id} align={column.align}>
-                        {editing ? (
+                        {editing && row.id === rowEditing ? (
                           <Input
                             // type={String}
                             defaultValue={value}
                             name={column.id}
-                            onChange={e => onChange(e, row)} />
+                            onChange={(e) => onChange(e, row)}
+                          />
+                        ) : column.format && typeof value === 'number' ? (
+                          column.format(value)
                         ) : (
-                          column.format && typeof value === 'number'
-                            ? column.format(value)
-                            : value
+                          value
                         )}
                       </TableCell>
                     );
                   })}
-                  {editing ? (
+                  {editing && row.id === rowEditing ? (
                     <DoneIcon onClick={() => onDone(row)} />
-                  ) :
+                  ) : (
                     <>
                       <TableCell>
-                        <EditIcon onClick={() => onEdit()} />
+                        <EditIcon onClick={() => onEdit(row)} />
                       </TableCell>
                       <TableCell>
                         <DeleteIcon onClick={() => onDelete(row)} />
                       </TableCell>
                     </>
-                  }
-
+                  )}
                 </TableRow>
               ))}
           </TableBody>
